First run

+++ trunk/src/test/java/home/mutant/bayes/NaiveBayesTest.java
@@ -13,8 +13,8 @@ public class NaiveBayesTest
 	public void testPosterior()
 	{
 		NaiveBayes bayes = new NaiveBayes();
-		List<Short> features = new ArrayList<Short>();
-		features.add((short)33);
+		List<Integer> features = new ArrayList<Integer>();
+		features.add(33);
 		bayes.addClassSample(features);
 		assertEquals(bayes.getPosterior(features), 0.7999975586012006,0.00000001);
 		bayes.addNonClassSample(features);
@@ -25,8 +25,8 @@ public class NaiveBayesTest
 	public void testPosteriorNonClass()
 	{
 		NaiveBayes bayes = new NaiveBayes();
-		List<Short> features = new ArrayList<Short>();
-		features.add((short)33);
+		List<Integer> features = new ArrayList<Integer>();
+		features.add(33);
 		bayes.addNonClassSample(features);
 		assertEquals(bayes.getPosterior(features), 1-0.7999975586012006,0.00000001);
 	}
@@ -35,9 +35,9 @@ public class NaiveBayesTest
 	public void testPosteriorClassOneNoise()
 	{
 		NaiveBayes bayes = new NaiveBayes();
-		List<Short> features = new ArrayList<Short>();
-		features.add((short)33);
-		features.add((short)44);
+		List<Integer> features = new ArrayList<Integer>();
+		features.add(33);
+		features.add(44);
 		bayes.addClassSample(features);
 		
 		features.remove(1);
@@ -47,34 +47,33 @@ public class NaiveBayesTest
 	public void testPosteriorClassManyNoise()
 	{
 		NaiveBayes bayes = new NaiveBayes();
-		List<Short> features = new ArrayList<Short>();
-		features.add((short)33);
-		features.add((short)11);
-		features.add((short)442);
-		features.add((short)3333);
-		features.add((short)456);
-		features.add((short)44);
+		List<Integer> features = new ArrayList<Integer>();
+		features.add(33);
+		features.add(11);
+		features.add(442);
+		features.add(3333);
+		features.add(456);
+		features.add(44);
 		
 		bayes.addClassSample(features);
 		
 		features.clear();
-		features.add((short)33);
+		features.add(33);
 		assertEquals(bayes.getPosterior(features), 0.799985351,0.00000001);
 	}
 	@Test
 	public void testPosteriorManySamplesOneFeature()
 	{
 		NaiveBayes bayes = new NaiveBayes();
-		List<Short> features = new ArrayList<Short>();
-		features.add((short)33);
+		List<Integer> features = new ArrayList<Integer>();
+		features.add(33);
 		bayes.addClassSample(features);
-		features.add((short)33);
+		features.add(33);
 		bayes.addClassSample(features);
-		features.add((short)33);
+		features.add(33);
 		bayes.addClassSample(features);
-		features.add((short)33);
+		features.add(33);
 		bayes.addClassSample(features);
 		assertEquals(bayes.getPosterior(features), 0.99998633,0.00000001);
-
 	}
 }
